send events from server to client
