+++ app/models/map.rb
@@ -33,6 +33,7 @@ class Map < ApplicationRecord
   # Validate the attached image is image/jpg, image/png, etc
   validates_attachment_content_type :screenshot, content_type: /\Aimage\/.*\Z/
 
+  after_update :after_updated
   after_save :update_deferring_topics_and_synapses, if: :permission_changed?
 
   delegate :count, to: :topics, prefix: :topic # same as `def topic_count; topics.count; end`
@@ -118,6 +119,13 @@ class Map < ApplicationRecord
       old_user_id
     end
     removed.compact
+  end
+  
+  def after_updated
+    attrs = ['name', 'desc', 'permission']
+    if attrs.any? {|k| changed_attributes.key?(k)}
+      ActionCable.server.broadcast 'map_' + id.to_s, type: 'mapUpdated'
+    end
   end
 
   def update_deferring_topics_and_synapses
