+++ app/models/message.rb
@@ -4,6 +4,8 @@ class Message < ApplicationRecord
   belongs_to :resource, polymorphic: true
 
   delegate :name, to: :user, prefix: true
+  
+  after_create :after_created
 
   def user_image
     user.image.url
@@ -12,5 +14,9 @@ class Message < ApplicationRecord
   def as_json(_options = {})
     json = super(methods: [:user_name, :user_image])
     json
+  end
+  
+  def after_created
+    ActionCable.server.broadcast 'map_' + map.id.to_s, type: 'messageCreated', message: self.as_json
   end
 end
