+++ frontend/src/Metamaps/Realtime/events.js
@@ -33,13 +33,11 @@ module.exports = {
 
   /* EVENTS RECEIVABLE FROM RAILS SERVER THROUGH ACTIONCABLE */
   MESSAGE_CREATED: 'MESSAGE_CREATED',
-  TOPIC_CREATED: 'TOPIC_CREATED',
+  TOPIC_ADDED: 'TOPIC_ADDED',
   TOPIC_UPDATED: 'TOPIC_UPDATED',
   TOPIC_REMOVED: 'TOPIC_REMOVED',
-  TOPIC_DELETED: 'TOPIC_DELETED',
-  SYNAPSE_CREATED: 'SYNAPSE_CREATED',
+  SYNAPSE_ADDED: 'SYNAPSE_ADDED',
   SYNAPSE_UPDATED: 'SYNAPSE_UPDATED',
   SYNAPSE_REMOVED: 'SYNAPSE_REMOVED',
-  SYNAPSE_DELETED: 'SYNAPSE_DELETED',
   MAP_UPDATED: 'MAP_UPDATED'
 }
