+++ frontend/src/Metamaps/Realtime/receivable.js
@@ -13,7 +13,6 @@ import { ChatView } from '../Views'
 import DataModel from '../DataModel'
 import GlobalUI from '../GlobalUI'
 import Control from '../Control'
-import Map from '../Map'
 import Mapper from '../Mapper'
 import Topic from '../Topic'
 import Synapse from '../Synapse'
@@ -25,188 +24,8 @@ export const juntoUpdated = self => state => {
   $(document).trigger(JUNTO_UPDATED)
 }
 
-export const synapseRemoved = self => data => {
-  var synapse = DataModel.Synapses.get(data.mappableid)
-  if (synapse) {
-    var edge = synapse.get('edge')
-    var mapping = synapse.getMapping()
-    if (edge.getData('mappings').length - 1 === 0) {
-      Control.hideEdge(edge)
-    }
-
-    var index = indexOf(edge.getData('synapses'), synapse)
-    edge.getData('mappings').splice(index, 1)
-    edge.getData('synapses').splice(index, 1)
-    if (edge.getData('displayIndex')) {
-      delete edge.data.$displayIndex
-    }
-    DataModel.Synapses.remove(synapse)
-    DataModel.Mappings.remove(mapping)
-  }
-}
-
-export const synapseDeleted = self => data => {
-  synapseRemoved(self)(data)
-}
-
-export const synapseCreated = self => data => {
-  var topic1, topic2, node1, node2, synapse, mapping, cancel, mapper
-
-  function waitThenRenderSynapse() {
-    if (synapse && mapping && mapper) {
-      topic1 = synapse.getTopic1()
-      node1 = topic1.get('node')
-      topic2 = synapse.getTopic2()
-      node2 = topic2.get('node')
-
-      Synapse.renderSynapse(mapping, synapse, node1, node2, false)
-    } else if (!cancel) {
-      setTimeout(waitThenRenderSynapse, 10)
-    }
-  }
-
-  mapper = DataModel.Mappers.get(data.mapperid)
-  if (mapper === undefined) {
-    Mapper.get(data.mapperid, function(m) {
-      DataModel.Mappers.add(m)
-      mapper = m
-    })
-  }
-  $.ajax({
-    url: '/synapses/' + data.mappableid + '.json',
-    success: function(response) {
-      DataModel.Synapses.add(response)
-      synapse = DataModel.Synapses.get(response.id)
-    },
-    error: function() {
-      cancel = true
-    }
-  })
-  $.ajax({
-    url: '/mappings/' + data.mappingid + '.json',
-    success: function(response) {
-      DataModel.Mappings.add(response)
-      mapping = DataModel.Mappings.get(response.id)
-    },
-    error: function() {
-      cancel = true
-    }
-  })
-  waitThenRenderSynapse()
-}
-
-export const topicRemoved = self => data => {
-  var topic = DataModel.Topics.get(data.mappableid)
-  if (topic) {
-    var node = topic.get('node')
-    var mapping = topic.getMapping()
-    Control.hideNode(node.id)
-    DataModel.Topics.remove(topic)
-    DataModel.Mappings.remove(mapping)
-  }
-}
-
-export const topicDeleted = self => data => {
-  topicRemoved(self)(data)
-}
-
-export const topicCreated = self => data => {
-  var topic, mapping, mapper, cancel
-
-  function waitThenRenderTopic() {
-    if (topic && mapping && mapper) {
-      Topic.renderTopic(mapping, topic, false, false)
-    } else if (!cancel) {
-      setTimeout(waitThenRenderTopic, 10)
-    }
-  }
-
-  mapper = DataModel.Mappers.get(data.mapperid)
-  if (mapper === undefined) {
-    Mapper.get(data.mapperid, function(m) {
-      DataModel.Mappers.add(m)
-      mapper = m
-    })
-  }
-  $.ajax({
-    url: '/topics/' + data.mappableid + '.json',
-    success: function(response) {
-      DataModel.Topics.add(response)
-      topic = DataModel.Topics.get(response.id)
-    },
-    error: function() {
-      cancel = true
-    }
-  })
-  $.ajax({
-    url: '/mappings/' + data.mappingid + '.json',
-    success: function(response) {
-      DataModel.Mappings.add(response)
-      mapping = DataModel.Mappings.get(response.id)
-    },
-    error: function() {
-      cancel = true
-    }
-  })
-
-  waitThenRenderTopic()
-}
-
-export const messageCreated = self => data => {
-  ChatView.addMessages(new DataModel.MessageCollection(data))
-}
-
-export const mapUpdated = self => data => {
-  var map = Active.Map
-  var isActiveMap = map && data.mapId === map.id
-  if (isActiveMap) {
-    var couldEditBefore = map.authorizeToEdit(Active.Mapper)
-    var idBefore = map.id
-    map.fetch({
-      success: function(model, response) {
-        var idNow = model.id
-        var canEditNow = model.authorizeToEdit(Active.Mapper)
-        if (idNow !== idBefore) {
-          Map.leavePrivateMap() // this means the map has been changed to private
-        } else if (couldEditBefore && !canEditNow) {
-          Map.cantEditNow()
-        } else if (!couldEditBefore && canEditNow) {
-          Map.canEditNow()
-        } else {
-          model.trigger('changeByOther')
-        }
-      }
-    })
-  }
-}
-
-export const topicUpdated = self => data => {
-  var topic = DataModel.Topics.get(data.topicId)
-  if (topic) {
-    var node = topic.get('node')
-    topic.fetch({
-      success: function(model) {
-        model.set({ node: node })
-        model.trigger('changeByOther')
-      }
-    })
-  }
-}
-
-export const synapseUpdated = self => data => {
-  var synapse = DataModel.Synapses.get(data.synapseId)
-  if (synapse) {
-    // edge reset necessary because fetch causes model reset
-    var edge = synapse.get('edge')
-    synapse.fetch({
-      success: function(model) {
-        model.set({ edge: edge })
-        model.trigger('changeByOther')
-      }
-    })
-  }
-}
-
+/* All the following events are received through the nodejs realtime server
+    and are done this way because they are transient data, not persisted to the server */
 export const topicDragged = self => positions => {
   var topic
   var node
